# docs/_제출본/openapi.yaml
openapi: 3.0.3
info:
  title: 오늘의직관 API
  description: |
    KBO 직관 기록 및 관람 정보 서비스 REST API 명세서.

    - 작성자: 이채목
    - 작성일: 2026-08-18

    ### 데이터 출처
    경기 데이터는 이용이 명시적으로 허용된 공식 API(TheSportsDB, Korean KBO League ID 4830)에서
    배치로 동기화하며, 누락·불일치는 사용자 제보로 보완한다 (REQ-F-107, REQ-N-013).
    외부 데이터가 포함된 응답에는 `dataSource` 필드로 출처를 표기한다 (REQ-N-014).
    클라이언트 요청이 외부 API로 직접 전달되지 않는다 (REQ-N-016).

    ### 공통 규칙
    - 인증이 필요한 API는 `Authorization: Bearer {accessToken}` 헤더를 요구한다.
    - 모든 오류 응답은 `ErrorResponse` 형식을 따른다.
    - 목록 조회는 `page`, `size` 쿼리 파라미터를 사용한다 (page는 0부터).
  version: 0.1.0
  contact:
    name: 이채목
servers:
  - url: https://api.todayjikgwan.com/api/v1
    description: 운영 서버
  - url: http://localhost:8080/api/v1
    description: 로컬 개발 서버
...
security:
  - bearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: 이메일 회원가입
      description: REQ-F-001. 가입 후 인증 메일을 발송한다. 인증 링크 유효기간은 30분이다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, nickname]
              properties:
                email:
                  type: string
                  format: email
                  example: fan@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: 8자 이상, 영문/숫자/특수문자 중 2종 이상
                nickname:
                  type: string
                  maxLength: 30
                  example: 잠실단골
                favoriteTeamId:
                  type: integer
                  format: int64
                  nullable: true
      responses:
        '201':
          description: 가입 성공. 인증 메일 발송됨
        '409':
          description: 이메일 또는 닉네임 중복
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                code: DUPLICATE_EMAIL
                message: 이미 사용 중인 이메일입니다.
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/login:
    post:
      tags: [Auth]
      summary: 이메일 로그인
      description: REQ-F-002. 비밀번호 5회 연속 실패 시 캡차 토큰을 요구한다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                captchaToken:
                  type: string
                  nullable: true
                  description: 5회 이상 실패한 계정에서만 필수
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: 계정 정보 불일치. 이메일/비밀번호 중 무엇이 틀렸는지 구분하지 않는다
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                code: INVALID_CREDENTIALS
                message: 이메일 또는 비밀번호가 올바르지 않습니다.
        '403':
          description: 이메일 미인증 또는 캡차 필요
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                code: CAPTCHA_REQUIRED
                message: 캡차 인증이 필요합니다.
...
  /stadiums/{stadiumId}:
    get:
      tags: [Stadium]
      summary: 구장 상세 조회
      description: REQ-F-109, REQ-F-110. 구역별 만족도 집계를 함께 반환한다.
      security: []
      parameters:
        - { name: stadiumId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StadiumDetailResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /stadiums/{stadiumId}/zones/{zoneId}/reviews:
    get:
      tags: [Stadium]
      summary: 구역별 후기 조회
      description: REQ-F-111. 공개 설정된 기록의 메모만 최신순으로 반환하며 작성자는 노출하지 않는다.
      security: []
      parameters:
        - { name: stadiumId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: zoneId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    memo: { type: string }
                    rating: { type: integer, example: 4 }
                    attendedAt: { type: string, format: date }

  # ==================== AttendanceLog ====================
  /attendance-logs:
    get:
      tags: [AttendanceLog]
      summary: 내 직관 기록 목록
      description: REQ-F-213
      parameters:
        - { name: season, in: query, schema: { type: integer } }
        - { name: stadiumId, in: query, schema: { type: integer, format: int64 } }
        - { name: teamId, in: query, schema: { type: integer, format: int64 } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AttendanceLogSummary' }
    post:
      tags: [AttendanceLog]
      summary: 직관 기록 작성
      description: |
        REQ-F-201 ~ 211. 저장 성공 시 전적 재집계를 트리거한다 (REQ-F-309).

        **동일 경기에 대한 기록은 사용자당 1건으로 제한된다.** 이미 존재하면 409를 반환하며,
        클라이언트는 기존 기록 수정으로 유도한다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendanceLogRequest' }
      responses:
        '201':
          description: 작성 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceLogDetail' }
        '409':
          description: 동일 경기 기록이 이미 존재함
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                code: DUPLICATE_ATTENDANCE_LOG
                message: 이미 이 경기 기록이 있어요.
                data: { existingLogId: 1024 }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /attendance-logs/{logId}/photos:
    post:
      tags: [AttendanceLog]
      summary: 사진 업로드
      description: |
        REQ-F-203. 기록당 최대 10장, 1장당 10MB 이하.
        서버는 저장 전에 **위치 메타데이터를 제거하고**(REQ-N-007) 썸네일을 생성한다(REQ-N-002).
        일부 파일이 실패해도 나머지는 정상 처리하며, 실패 목록을 응답에 포함한다.
      parameters:
        - { $ref: '#/components/parameters/LogId' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        '201':
          description: 업로드 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploaded:
                    type: array
                    items: { $ref: '#/components/schemas/PhotoResponse' }
                  failed:
                    type: array
                    items:
                      type: object
                      properties:
                        fileName: { type: string }
                        reason: { type: string, example: FILE_TOO_LARGE }
        '413':
          description: 전체 파일이 허용 용량을 초과함

  # ==================== Stat ====================
...
  /companion-posts:
    get:
      tags: [Companion]
      summary: 동행 모집 목록
      description: REQ-F-502. 마감된 모집은 기본적으로 제외한다.
      parameters:
        - { name: gameDate, in: query, schema: { type: string, format: date } }
        - { name: stadiumId, in: query, schema: { type: integer, format: int64 } }
        - { name: includeClosed, in: query, schema: { type: boolean, default: false } }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CompanionPostResponse' }
    post:
      tags: [Companion]
      summary: 동행 모집 등록
      description: REQ-F-501. 이미 시작한 경기에는 등록할 수 없다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gameId, capacity]
              properties:
                gameId: { type: integer, format: int64 }
                capacity: { type: integer, minimum: 2, maximum: 10, description: '작성자 포함 정원' }
                intro: { type: string, maxLength: 500 }
      responses:
        '201':
          description: 등록 성공
        '400':
          description: 경기 시작 시각이 지난 경우
  /companion-posts/{postId}:
    get:
      tags: [Companion]
      summary: 동행 모집 상세
      parameters:
        - { $ref: '#/components/parameters/PostId' }
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompanionPostDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /companion-posts/{postId}/applications:
    post:
      tags: [Companion]
      summary: 동행 참여 (선착순 확정)
      description: |
        REQ-F-503, REQ-F-504, REQ-F-505, REQ-N-011.

        **동시 요청 시에도 확정 인원이 정원을 초과하지 않아야 한다.**
        서버는 게시글 행에 대한 잠금 또는 낙관적 락(version)으로 정합성을 보장하며,
        정원이 이미 찬 요청에는 409를 반환한다. 대기자 개념은 두지 않는다.
      parameters:
        - { $ref: '#/components/parameters/PostId' }
      responses:
        '201':
          description: 확정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  seq: { type: integer, description: '확정 순번', example: 3 }
                  confirmedCount: { type: integer, example: 3 }
                  capacity: { type: integer, example: 3 }
        '409':
          description: 정원 마감 또는 동일 시간대 중복 확정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                full:
                  summary: 정원 마감 (REQ-F-504)
                  value:
                    code: COMPANION_POST_FULL
                    message: 방금 정원이 찼어요.
                conflict:
                  summary: 동일 시간대 중복 (REQ-F-505)
                  value:
                    code: TIME_CONFLICT
                    message: 같은 시간에 확정된 동행이 있어요.
        '400':
          description: 본인이 등록한 모집에는 지원할 수 없음

    delete:
      tags: [Companion]
      summary: 동행 참여 취소
      description: REQ-F-506. 경기 시작 24시간 이내에는 취소할 수 없다. 취소 시 해당 자리는 다시 모집 상태가 된다.
      parameters:
        - { $ref: '#/components/parameters/PostId' }
      responses:
        '204':
          description: 취소 성공
        '400':
          description: 취소 제한 시간 경과
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                code: CANCEL_DEADLINE_PASSED
                message: 경기 임박 시에는 취소할 수 없어요.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
    LogId:
      name: logId
      in: path
      required: true
      schema: { type: integer, format: int64 }
    PostId:
      name: postId
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: 요청 값이 올바르지 않음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: 대상을 찾을 수 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        code: { type: string, example: INVALID_CREDENTIALS }
        message: { type: string, example: 이메일 또는 비밀번호가 올바르지 않습니다. }
        data: { type: object, nullable: true }
    TokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, example: 3600 }
...
    TeamResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: 두산 베어스 }
        shortName: { type: string, example: 두산 }
        logoUrl: { type: string, nullable: true }
        channels: { $ref: '#/components/schemas/TeamChannels' }
    GameResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        seasonYear: { type: integer, example: 2026 }
        gameDate: { type: string, format: date }
        startAt: { type: string, format: date-time }
        stadium: { $ref: '#/components/schemas/StadiumResponse' }
        homeTeam: { $ref: '#/components/schemas/TeamResponse' }
        awayTeam: { $ref: '#/components/schemas/TeamResponse' }
        homeScore: { type: integer, nullable: true }
        awayScore: { type: integer, nullable: true }
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, FINISHED, CANCELED, SUSPENDED]
        resultConfirmed:
          type: boolean
          description: REQ-F-608. false 이면 화면에 "확인 중" 배지를 표시한다
          example: true
...
    TeamChannels:
      type: object
      description: REQ-F-114 구단 공식 채널
      properties:
        homepageUrl: { type: string, nullable: true }
        ticketUrl: { type: string, nullable: true }
        instagramUrl: { type: string, nullable: true }
        youtubeUrl: { type: string, nullable: true }
    StadiumResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: 잠실야구장 }
        address: { type: string }
        capacity: { type: integer, example: 25000 }
    StadiumDetailResponse:
      allOf:
        - $ref: '#/components/schemas/StadiumResponse'
        - type: object
          properties:
            homeTeams:
              type: array
              items: { $ref: '#/components/schemas/TeamResponse' }
            zones:
              type: array
              items: { $ref: '#/components/schemas/ZoneStatResponse' }
            myRecord:
              type: object
              nullable: true
              description: 로그인 사용자의 해당 구장 전적
              properties:
                games: { type: integer }
                wins: { type: integer }
                draws: { type: integer }
                losses: { type: integer }
    ZoneStatResponse:
      type: object
      properties:
        zoneId: { type: integer, format: int64 }
        name: { type: string, example: 3루 내야 지정석 }
        avgRating:
          type: number
          format: float
          nullable: true
          description: 표본 5건 미만이면 null (REQ-F-305)
          example: 4.2
        ratingCount: { type: integer, example: 87 }
        smallSample: { type: boolean, example: false }
    AttendanceLogRequest:
      type: object
      required: [gameId, stadiumZoneId, zoneRating]
      properties:
        gameId: { type: integer, format: int64 }
        cheerTeamId:
          type: integer
          format: int64
          nullable: true
          description: null 이면 중립 관람 (승패 집계 제외)
        stadiumZoneId: { type: integer, format: int64 }
        zoneRating: { type: integer, minimum: 1, maximum: 5, description: 필수 (REQ-F-206) }
        memo: { type: string, maxLength: 1000 }
        gameRating: { type: integer, minimum: 1, maximum: 5, nullable: true }
        ticketCost: { type: integer, minimum: 0, nullable: true }
        foodCost: { type: integer, minimum: 0, nullable: true }
        transportCost: { type: integer, minimum: 0, nullable: true }
        visibility: { type: string, enum: [PUBLIC, PRIVATE], default: PRIVATE }
        reportedHomeScore:
          type: integer
          minimum: 0
          nullable: true
          description: REQ-F-606. 경기 결과가 미확정(resultConfirmed=false)인 경우 필수
        reportedAwayScore:
          type: integer
          minimum: 0
          nullable: true
          description: REQ-F-606. 경기 결과가 미확정인 경우 필수
        companions:
          type: array
          maxItems: 10
          items:
            type: object
            properties:
              userId: { type: integer, format: int64, nullable: true }
              name: { type: string, nullable: true, description: 비회원인 경우 }
    AttendanceLogSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        gameDate: { type: string, format: date }
        stadiumName: { type: string }
        matchup: { type: string, example: '두산 5 : 3 롯데' }
        result: { type: string, enum: [WIN, DRAW, LOSE, NEUTRAL] }
        thumbnailUrl: { type: string, nullable: true }
    AttendanceLogDetail:
      allOf:
        - $ref: '#/components/schemas/AttendanceLogSummary'
        - type: object
          properties:
            game: { $ref: '#/components/schemas/GameResponse' }
            zoneName: { type: string }
            zoneRating: { type: integer }
            memo: { type: string, nullable: true }
            gameRating: { type: integer, nullable: true }
            totalCost: { type: integer, nullable: true }
            visibility: { type: string, enum: [PUBLIC, PRIVATE] }
            photos:
              type: array
              items: { $ref: '#/components/schemas/PhotoResponse' }
    PhotoResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        originalUrl: { type: string }
        thumbnailUrl: { type: string }
        takenAt: { type: string, format: date-time, nullable: true }
...
    CompanionPostResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        game: { $ref: '#/components/schemas/GameResponse' }
        authorNickname: { type: string }
        capacity: { type: integer }
        confirmedCount: { type: integer }
        status: { type: string, enum: [OPEN, FULL, CLOSED, ENDED] }
    CompanionPostDetail:
      allOf:
        - $ref: '#/components/schemas/CompanionPostResponse'
        - type: object
          properties:
            intro: { type: string, nullable: true }
            confirmedMembers:
              type: array
              items:
                type: object
                properties:
                  nickname: { type: string }
                  seq: { type: integer }
                  isAuthor: { type: boolean }
            myStatus:
              type: string
              enum: [NONE, CONFIRMED, AUTHOR]
              description: 클라이언트 버튼 상태 결정용
...